Quieting down a warning in Visual Studio

+++ src/parallel/ForceMatrixDecomposition.cpp
@@ -1448,9 +1448,9 @@ namespace OpenMD {
         }
         
         // find xyz-indices of cell that cutoffGroup is in.
-        whichCell.x() = nCells_.x() * scaled.x();
-        whichCell.y() = nCells_.y() * scaled.y();
-        whichCell.z() = nCells_.z() * scaled.z();
+        whichCell.x() = int(nCells_.x() * scaled.x());
+        whichCell.y() = int(nCells_.y() * scaled.y());
+        whichCell.z() = int(nCells_.z() * scaled.z());
         
         // find single index of this cell:
         cellIndex = Vlinear(whichCell, nCells_);
@@ -1476,9 +1476,9 @@ namespace OpenMD {
         }
         
         // find xyz-indices of cell that cutoffGroup is in.
-        whichCell.x() = nCells_.x() * scaled.x();
-        whichCell.y() = nCells_.y() * scaled.y();
-        whichCell.z() = nCells_.z() * scaled.z();
+        whichCell.x() = int(nCells_.x() * scaled.x());
+        whichCell.y() = int(nCells_.y() * scaled.y());
+        whichCell.z() = int(nCells_.z() * scaled.z());
         
         // find single index of this cell:
         cellIndex = Vlinear(whichCell, nCells_);
